Module perf_event_open_sys::bindings

Types and constants used with perf_event_open.

This module contains types and constants for use with the perf_event_open system call. These are automatically generated from the header files <linux/perf_event.h> and <linux/hw_breakpoint.h> by the Rust bindgen tool.

It’s not always obvious how bindgen will choose to reflect a given C construct into Rust. The best approach I’ve found is simply to search the source code for the C identifier name and see what bindgen did with it.

Structs

__BindgenBitfieldUnit

__IncompleteArrayField

__kernel_fd_set

__kernel_fsid_t

perf_branch_entry

perf_event_attr

perf_event_header

perf_event_mmap_page

perf_event_mmap_page__bindgen_ty_1__bindgen_ty_1

perf_event_query_bpf

perf_mem_data_src__bindgen_ty_1

perf_ns_link_info

Constants

CGROUP_NS_INDEX

HW_BREAKPOINT_EMPTY

HW_BREAKPOINT_INVALID

HW_BREAKPOINT_LEN_1

HW_BREAKPOINT_LEN_2

HW_BREAKPOINT_LEN_3

HW_BREAKPOINT_LEN_4

HW_BREAKPOINT_LEN_5

HW_BREAKPOINT_LEN_6

HW_BREAKPOINT_LEN_7

HW_BREAKPOINT_LEN_8

HW_BREAKPOINT_R

HW_BREAKPOINT_RW

HW_BREAKPOINT_W

HW_BREAKPOINT_X

IOCSIZE_MASK

IOCSIZE_SHIFT

IOC_IN

IOC_INOUT

IOC_OUT

IPC_NS_INDEX

MNT_NS_INDEX

NET_NS_INDEX

NR_NAMESPACES

PERF_ATTR_SIZE_VER0

PERF_ATTR_SIZE_VER1

PERF_ATTR_SIZE_VER2

PERF_ATTR_SIZE_VER3

PERF_ATTR_SIZE_VER4

PERF_ATTR_SIZE_VER5

PERF_ATTR_SIZE_VER6

PERF_AUX_FLAG_COLLISION

PERF_AUX_FLAG_OVERWRITE

PERF_AUX_FLAG_PARTIAL

PERF_AUX_FLAG_TRUNCATED

PERF_BR_CALL

PERF_BR_COND

PERF_BR_COND_CALL

PERF_BR_COND_RET

PERF_BR_IND

PERF_BR_IND_CALL

PERF_BR_MAX

PERF_BR_RET

PERF_BR_SYSCALL

PERF_BR_SYSRET

PERF_BR_UNCOND

PERF_BR_UNKNOWN

PERF_FLAG_FD_CLOEXEC

PERF_FLAG_FD_NO_GROUP

PERF_FLAG_FD_OUTPUT

PERF_FLAG_PID_CGROUP

PERF_MAX_CONTEXTS_PER_STACK

PERF_MAX_STACK_DEPTH

PERF_MEM_LOCK_LOCKED

PERF_MEM_LOCK_NA

PERF_MEM_LOCK_SHIFT

PERF_MEM_LVLNUM_ANY_CACHE

PERF_MEM_LVLNUM_L1

PERF_MEM_LVLNUM_L2

PERF_MEM_LVLNUM_L3

PERF_MEM_LVLNUM_L4

PERF_MEM_LVLNUM_LFB

PERF_MEM_LVLNUM_NA

PERF_MEM_LVLNUM_PMEM

PERF_MEM_LVLNUM_RAM

PERF_MEM_LVLNUM_SHIFT

PERF_MEM_LVL_HIT

PERF_MEM_LVL_IO

PERF_MEM_LVL_L1

PERF_MEM_LVL_L2

PERF_MEM_LVL_L3

PERF_MEM_LVL_LFB

PERF_MEM_LVL_LOC_RAM

PERF_MEM_LVL_MISS

PERF_MEM_LVL_NA

PERF_MEM_LVL_REM_CCE1

PERF_MEM_LVL_REM_CCE2

PERF_MEM_LVL_REM_RAM1

PERF_MEM_LVL_REM_RAM2

PERF_MEM_LVL_SHIFT

PERF_MEM_LVL_UNC

PERF_MEM_OP_EXEC

PERF_MEM_OP_LOAD

PERF_MEM_OP_NA

PERF_MEM_OP_PFETCH

PERF_MEM_OP_SHIFT

PERF_MEM_OP_STORE

PERF_MEM_REMOTE_REMOTE

PERF_MEM_REMOTE_SHIFT

PERF_MEM_SNOOPX_FWD

PERF_MEM_SNOOPX_SHIFT

PERF_MEM_SNOOP_HIT

PERF_MEM_SNOOP_HITM

PERF_MEM_SNOOP_MISS

PERF_MEM_SNOOP_NA

PERF_MEM_SNOOP_NONE

PERF_MEM_SNOOP_SHIFT

PERF_MEM_TLB_HIT

PERF_MEM_TLB_L1

PERF_MEM_TLB_L2

PERF_MEM_TLB_MISS

PERF_MEM_TLB_NA

PERF_MEM_TLB_OS

PERF_MEM_TLB_SHIFT

PERF_MEM_TLB_WK

PERF_RECORD_KSYMBOL_FLAGS_UNREGISTER

PERF_RECORD_MISC_COMM_EXEC

PERF_RECORD_MISC_CPUMODE_MASK

PERF_RECORD_MISC_CPUMODE_UNKNOWN

PERF_RECORD_MISC_EXACT_IP

PERF_RECORD_MISC_EXT_RESERVED

PERF_RECORD_MISC_FORK_EXEC

PERF_RECORD_MISC_GUEST_KERNEL

PERF_RECORD_MISC_GUEST_USER

PERF_RECORD_MISC_HYPERVISOR

PERF_RECORD_MISC_KERNEL

PERF_RECORD_MISC_MMAP_DATA

PERF_RECORD_MISC_PROC_MAP_PARSE_TIMEOUT

PERF_RECORD_MISC_SWITCH_OUT

PERF_RECORD_MISC_SWITCH_OUT_PREEMPT

PERF_RECORD_MISC_USER

PERF_TXN_ABORT_MASK

PERF_TXN_ABORT_SHIFT

PERF_TXN_ASYNC

PERF_TXN_CAPACITY_READ

PERF_TXN_CAPACITY_WRITE

PERF_TXN_CONFLICT

PERF_TXN_ELISION

PERF_TXN_MAX

PERF_TXN_RETRY

PERF_TXN_SYNC

PERF_TXN_TRANSACTION

PID_NS_INDEX

USER_NS_INDEX

UTS_NS_INDEX

_ASM_X86_UNISTD_64_H

_IOC_DIRBITS

_IOC_DIRMASK

_IOC_DIRSHIFT

_IOC_NONE

_IOC_NRBITS

_IOC_NRMASK

_IOC_NRSHIFT

_IOC_READ

_IOC_SIZEBITS

_IOC_SIZEMASK

_IOC_SIZESHIFT

_IOC_TYPEBITS

_IOC_TYPEMASK

_IOC_TYPESHIFT

_IOC_WRITE

__BITS_PER_LONG

__FD_SETSIZE

__LITTLE_ENDIAN

__NR__sysctl

__NR_accept

__NR_accept4

__NR_access

__NR_acct

__NR_add_key

__NR_adjtimex

__NR_afs_syscall

__NR_alarm

__NR_arch_prctl

__NR_bind

__NR_bpf

__NR_brk

__NR_capget

__NR_capset

__NR_chdir

__NR_chmod

__NR_chown

__NR_chroot

__NR_clock_adjtime

__NR_clock_getres

__NR_clock_gettime

__NR_clock_nanosleep

__NR_clock_settime

__NR_clone

__NR_clone3

__NR_close

__NR_connect

__NR_copy_file_range

__NR_creat

__NR_create_module

__NR_delete_module

__NR_dup

__NR_dup2

__NR_dup3

__NR_epoll_create

__NR_epoll_create1

__NR_epoll_ctl

__NR_epoll_ctl_old

__NR_epoll_pwait

__NR_epoll_wait

__NR_epoll_wait_old

__NR_eventfd

__NR_eventfd2

__NR_execve

__NR_execveat

__NR_exit

__NR_exit_group

__NR_faccessat

__NR_fadvise64

__NR_fallocate

__NR_fanotify_init

__NR_fanotify_mark

__NR_fchdir

__NR_fchmod

__NR_fchmodat

__NR_fchown

__NR_fchownat

__NR_fcntl

__NR_fdatasync

__NR_fgetxattr

__NR_finit_module

__NR_flistxattr

__NR_flock

__NR_fork

__NR_fremovexattr

__NR_fsconfig

__NR_fsetxattr

__NR_fsmount

__NR_fsopen

__NR_fspick

__NR_fstat

__NR_fstatfs

__NR_fsync

__NR_ftruncate

__NR_futex

__NR_futimesat

__NR_get_kernel_syms

__NR_get_mempolicy

__NR_get_robust_list

__NR_get_thread_area

__NR_getcpu

__NR_getcwd

__NR_getdents

__NR_getdents64

__NR_getegid

__NR_geteuid

__NR_getgid

__NR_getgroups

__NR_getitimer

__NR_getpeername

__NR_getpgid

__NR_getpgrp

__NR_getpid

__NR_getpmsg

__NR_getppid

__NR_getpriority

__NR_getrandom

__NR_getresgid

__NR_getresuid

__NR_getrlimit

__NR_getrusage

__NR_getsid

__NR_getsockname

__NR_getsockopt

__NR_gettid

__NR_gettimeofday

__NR_getuid

__NR_getxattr

__NR_init_module

__NR_inotify_add_watch

__NR_inotify_init

__NR_inotify_init1

__NR_inotify_rm_watch

__NR_io_cancel

__NR_io_destroy

__NR_io_getevents

__NR_io_pgetevents

__NR_io_setup

__NR_io_submit

__NR_io_uring_enter

__NR_io_uring_register

__NR_io_uring_setup

__NR_ioctl

__NR_ioperm

__NR_iopl

__NR_ioprio_get

__NR_ioprio_set

__NR_kcmp

__NR_kexec_file_load

__NR_kexec_load

__NR_keyctl

__NR_kill

__NR_lchown

__NR_lgetxattr

__NR_link

__NR_linkat

__NR_listen

__NR_listxattr

__NR_llistxattr

__NR_lookup_dcookie

__NR_lremovexattr

__NR_lseek

__NR_lsetxattr

__NR_lstat

__NR_madvise

__NR_mbind

__NR_membarrier

__NR_memfd_create

__NR_migrate_pages

__NR_mincore

__NR_mkdir

__NR_mkdirat

__NR_mknod

__NR_mknodat

__NR_mlock

__NR_mlock2

__NR_mlockall

__NR_mmap

__NR_modify_ldt

__NR_mount

__NR_move_mount

__NR_move_pages

__NR_mprotect

__NR_mq_getsetattr

__NR_mq_notify

__NR_mq_open

__NR_mq_timedreceive

__NR_mq_timedsend

__NR_mq_unlink

__NR_mremap

__NR_msgctl

__NR_msgget

__NR_msgrcv

__NR_msgsnd

__NR_msync

__NR_munlock

__NR_munlockall

__NR_munmap

__NR_name_to_handle_at

__NR_nanosleep

__NR_newfstatat

__NR_nfsservctl

__NR_open

__NR_open_by_handle_at

__NR_open_tree

__NR_openat

__NR_openat2

__NR_pause

__NR_perf_event_open

__NR_personality

__NR_pidfd_getfd

__NR_pidfd_open

__NR_pidfd_send_signal

__NR_pipe

__NR_pipe2

__NR_pivot_root

__NR_pkey_alloc

__NR_pkey_free

__NR_pkey_mprotect

__NR_poll

__NR_ppoll

__NR_prctl

__NR_pread64

__NR_preadv

__NR_preadv2

__NR_prlimit64

__NR_process_vm_readv

__NR_process_vm_writev

__NR_pselect6

__NR_ptrace

__NR_putpmsg

__NR_pwrite64

__NR_pwritev

__NR_pwritev2

__NR_query_module

__NR_quotactl

__NR_read

__NR_readahead

__NR_readlink

__NR_readlinkat

__NR_readv

__NR_reboot

__NR_recvfrom

__NR_recvmmsg

__NR_recvmsg

__NR_remap_file_pages

__NR_removexattr

__NR_rename

__NR_renameat

__NR_renameat2

__NR_request_key

__NR_restart_syscall

__NR_rmdir

__NR_rseq

__NR_rt_sigaction

__NR_rt_sigpending

__NR_rt_sigprocmask

__NR_rt_sigqueueinfo

__NR_rt_sigreturn

__NR_rt_sigsuspend

__NR_rt_sigtimedwait

__NR_rt_tgsigqueueinfo

__NR_sched_get_priority_max

__NR_sched_get_priority_min

__NR_sched_getaffinity

__NR_sched_getattr

__NR_sched_getparam

__NR_sched_getscheduler

__NR_sched_rr_get_interval

__NR_sched_setaffinity

__NR_sched_setattr

__NR_sched_setparam

__NR_sched_setscheduler

__NR_sched_yield

__NR_seccomp

__NR_security

__NR_select

__NR_semctl

__NR_semget

__NR_semop

__NR_semtimedop

__NR_sendfile

__NR_sendmmsg

__NR_sendmsg

__NR_sendto

__NR_set_mempolicy

__NR_set_robust_list

__NR_set_thread_area

__NR_set_tid_address

__NR_setdomainname

__NR_setfsgid

__NR_setfsuid

__NR_setgid

__NR_setgroups

__NR_sethostname

__NR_setitimer

__NR_setns

__NR_setpgid

__NR_setpriority

__NR_setregid

__NR_setresgid

__NR_setresuid

__NR_setreuid

__NR_setrlimit

__NR_setsid

__NR_setsockopt

__NR_settimeofday

__NR_setuid

__NR_setxattr

__NR_shmat

__NR_shmctl

__NR_shmdt

__NR_shmget

__NR_shutdown

__NR_sigaltstack

__NR_signalfd

__NR_signalfd4

__NR_socket

__NR_socketpair

__NR_splice

__NR_stat

__NR_statfs

__NR_statx

__NR_swapoff

__NR_swapon

__NR_symlink

__NR_symlinkat

__NR_sync

__NR_sync_file_range

__NR_syncfs

__NR_sysfs

__NR_sysinfo

__NR_syslog

__NR_tee

__NR_tgkill

__NR_time

__NR_timer_create

__NR_timer_delete

__NR_timer_getoverrun

__NR_timer_gettime

__NR_timer_settime

__NR_timerfd_create

__NR_timerfd_gettime

__NR_timerfd_settime

__NR_times

__NR_tkill

__NR_truncate

__NR_tuxcall

__NR_umask

__NR_umount2

__NR_uname

__NR_unlink

__NR_unlinkat

__NR_unshare

__NR_uselib

__NR_userfaultfd

__NR_ustat

__NR_utime

__NR_utimensat

__NR_utimes

__NR_vfork

__NR_vhangup

__NR_vmsplice

__NR_vserver

__NR_wait4

__NR_waitid

__NR_write

__NR_writev

__X32_SYSCALL_BIT

bp_type_idx_TYPE_DATA

bp_type_idx_TYPE_INST

bp_type_idx_TYPE_MAX

perf_bpf_event_type_PERF_BPF_EVENT_MAX

perf_bpf_event_type_PERF_BPF_EVENT_PROG_LOAD

perf_bpf_event_type_PERF_BPF_EVENT_PROG_UNLOAD

perf_bpf_event_type_PERF_BPF_EVENT_UNKNOWN

perf_branch_sample_type_PERF_SAMPLE_BRANCH_ABORT_TX

perf_branch_sample_type_PERF_SAMPLE_BRANCH_ANY

perf_branch_sample_type_PERF_SAMPLE_BRANCH_ANY_CALL

perf_branch_sample_type_PERF_SAMPLE_BRANCH_ANY_RETURN

perf_branch_sample_type_PERF_SAMPLE_BRANCH_CALL

perf_branch_sample_type_PERF_SAMPLE_BRANCH_CALL_STACK

perf_branch_sample_type_PERF_SAMPLE_BRANCH_COND

perf_branch_sample_type_PERF_SAMPLE_BRANCH_HV

perf_branch_sample_type_PERF_SAMPLE_BRANCH_IND_CALL

perf_branch_sample_type_PERF_SAMPLE_BRANCH_IND_JUMP

perf_branch_sample_type_PERF_SAMPLE_BRANCH_IN_TX

perf_branch_sample_type_PERF_SAMPLE_BRANCH_KERNEL

perf_branch_sample_type_PERF_SAMPLE_BRANCH_MAX

perf_branch_sample_type_PERF_SAMPLE_BRANCH_NO_CYCLES

perf_branch_sample_type_PERF_SAMPLE_BRANCH_NO_FLAGS

perf_branch_sample_type_PERF_SAMPLE_BRANCH_NO_TX

perf_branch_sample_type_PERF_SAMPLE_BRANCH_TYPE_SAVE

perf_branch_sample_type_PERF_SAMPLE_BRANCH_USER

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_ABORT_TX_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_ANY_CALL_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_ANY_RETURN_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_ANY_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_CALL_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_CALL_STACK_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_COND_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_HV_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_IND_CALL_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_IND_JUMP_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_IN_TX_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_KERNEL_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_MAX_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_NO_CYCLES_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_NO_FLAGS_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_NO_TX_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_TYPE_SAVE_SHIFT

perf_branch_sample_type_shift_PERF_SAMPLE_BRANCH_USER_SHIFT

perf_callchain_context_PERF_CONTEXT_GUEST

perf_callchain_context_PERF_CONTEXT_GUEST_KERNEL

perf_callchain_context_PERF_CONTEXT_GUEST_USER

perf_callchain_context_PERF_CONTEXT_HV

perf_callchain_context_PERF_CONTEXT_KERNEL

perf_callchain_context_PERF_CONTEXT_MAX

perf_callchain_context_PERF_CONTEXT_USER

perf_event_ioc_flags_PERF_IOC_FLAG_GROUP

perf_event_ioctls_DISABLE

perf_event_ioctls_ENABLE

perf_event_ioctls_ID

perf_event_ioctls_MODIFY_ATTRIBUTES

perf_event_ioctls_PAUSE_OUTPUT

perf_event_ioctls_PERIOD

perf_event_ioctls_QUERY_BPF

perf_event_ioctls_REFRESH

perf_event_ioctls_RESET

perf_event_ioctls_SET_BPF

perf_event_ioctls_SET_FILTER

perf_event_ioctls_SET_OUTPUT

perf_event_read_format_PERF_FORMAT_GROUP

perf_event_read_format_PERF_FORMAT_ID

perf_event_read_format_PERF_FORMAT_MAX

perf_event_read_format_PERF_FORMAT_TOTAL_TIME_ENABLED

perf_event_read_format_PERF_FORMAT_TOTAL_TIME_RUNNING

perf_event_sample_format_PERF_SAMPLE_ADDR

perf_event_sample_format_PERF_SAMPLE_AUX

perf_event_sample_format_PERF_SAMPLE_BRANCH_STACK

perf_event_sample_format_PERF_SAMPLE_CALLCHAIN

perf_event_sample_format_PERF_SAMPLE_CPU

perf_event_sample_format_PERF_SAMPLE_DATA_SRC

perf_event_sample_format_PERF_SAMPLE_ID

perf_event_sample_format_PERF_SAMPLE_IDENTIFIER

perf_event_sample_format_PERF_SAMPLE_IP

perf_event_sample_format_PERF_SAMPLE_MAX

perf_event_sample_format_PERF_SAMPLE_PERIOD

perf_event_sample_format_PERF_SAMPLE_PHYS_ADDR

perf_event_sample_format_PERF_SAMPLE_RAW

perf_event_sample_format_PERF_SAMPLE_READ

perf_event_sample_format_PERF_SAMPLE_REGS_INTR

perf_event_sample_format_PERF_SAMPLE_REGS_USER

perf_event_sample_format_PERF_SAMPLE_STACK_USER

perf_event_sample_format_PERF_SAMPLE_STREAM_ID

perf_event_sample_format_PERF_SAMPLE_TID

perf_event_sample_format_PERF_SAMPLE_TIME

perf_event_sample_format_PERF_SAMPLE_TRANSACTION

perf_event_sample_format_PERF_SAMPLE_WEIGHT

perf_event_sample_format___PERF_SAMPLE_CALLCHAIN_EARLY

perf_event_type_PERF_RECORD_AUX

perf_event_type_PERF_RECORD_BPF_EVENT

perf_event_type_PERF_RECORD_COMM

perf_event_type_PERF_RECORD_EXIT

perf_event_type_PERF_RECORD_FORK

perf_event_type_PERF_RECORD_ITRACE_START

perf_event_type_PERF_RECORD_KSYMBOL

perf_event_type_PERF_RECORD_LOST

perf_event_type_PERF_RECORD_LOST_SAMPLES

perf_event_type_PERF_RECORD_MAX

perf_event_type_PERF_RECORD_MMAP

perf_event_type_PERF_RECORD_MMAP2

perf_event_type_PERF_RECORD_NAMESPACES

perf_event_type_PERF_RECORD_READ

perf_event_type_PERF_RECORD_SAMPLE

perf_event_type_PERF_RECORD_SWITCH

perf_event_type_PERF_RECORD_SWITCH_CPU_WIDE

perf_event_type_PERF_RECORD_THROTTLE

perf_event_type_PERF_RECORD_UNTHROTTLE

perf_hw_cache_id_PERF_COUNT_HW_CACHE_BPU

perf_hw_cache_id_PERF_COUNT_HW_CACHE_DTLB

perf_hw_cache_id_PERF_COUNT_HW_CACHE_ITLB

perf_hw_cache_id_PERF_COUNT_HW_CACHE_L1D

perf_hw_cache_id_PERF_COUNT_HW_CACHE_L1I

perf_hw_cache_id_PERF_COUNT_HW_CACHE_LL

perf_hw_cache_id_PERF_COUNT_HW_CACHE_MAX

perf_hw_cache_id_PERF_COUNT_HW_CACHE_NODE

perf_hw_cache_op_id_PERF_COUNT_HW_CACHE_OP_MAX

perf_hw_cache_op_id_PERF_COUNT_HW_CACHE_OP_PREFETCH

perf_hw_cache_op_id_PERF_COUNT_HW_CACHE_OP_READ

perf_hw_cache_op_id_PERF_COUNT_HW_CACHE_OP_WRITE

perf_hw_cache_op_result_id_PERF_COUNT_HW_CACHE_RESULT_ACCESS

perf_hw_cache_op_result_id_PERF_COUNT_HW_CACHE_RESULT_MAX

perf_hw_cache_op_result_id_PERF_COUNT_HW_CACHE_RESULT_MISS

perf_hw_id_PERF_COUNT_HW_BRANCH_INSTRUCTIONS

perf_hw_id_PERF_COUNT_HW_BRANCH_MISSES

perf_hw_id_PERF_COUNT_HW_BUS_CYCLES

perf_hw_id_PERF_COUNT_HW_CACHE_MISSES

perf_hw_id_PERF_COUNT_HW_CACHE_REFERENCES

perf_hw_id_PERF_COUNT_HW_CPU_CYCLES

perf_hw_id_PERF_COUNT_HW_INSTRUCTIONS

perf_hw_id_PERF_COUNT_HW_MAX

perf_hw_id_PERF_COUNT_HW_REF_CPU_CYCLES

perf_hw_id_PERF_COUNT_HW_STALLED_CYCLES_BACKEND

perf_hw_id_PERF_COUNT_HW_STALLED_CYCLES_FRONTEND

perf_record_ksymbol_type_PERF_RECORD_KSYMBOL_TYPE_BPF

perf_record_ksymbol_type_PERF_RECORD_KSYMBOL_TYPE_MAX

perf_record_ksymbol_type_PERF_RECORD_KSYMBOL_TYPE_UNKNOWN

perf_sample_regs_abi_PERF_SAMPLE_REGS_ABI_32

perf_sample_regs_abi_PERF_SAMPLE_REGS_ABI_64

perf_sample_regs_abi_PERF_SAMPLE_REGS_ABI_NONE

perf_sw_ids_PERF_COUNT_SW_ALIGNMENT_FAULTS

perf_sw_ids_PERF_COUNT_SW_BPF_OUTPUT

perf_sw_ids_PERF_COUNT_SW_CONTEXT_SWITCHES

perf_sw_ids_PERF_COUNT_SW_CPU_CLOCK

perf_sw_ids_PERF_COUNT_SW_CPU_MIGRATIONS

perf_sw_ids_PERF_COUNT_SW_DUMMY

perf_sw_ids_PERF_COUNT_SW_EMULATION_FAULTS

perf_sw_ids_PERF_COUNT_SW_MAX

perf_sw_ids_PERF_COUNT_SW_PAGE_FAULTS

perf_sw_ids_PERF_COUNT_SW_PAGE_FAULTS_MAJ

perf_sw_ids_PERF_COUNT_SW_PAGE_FAULTS_MIN

perf_sw_ids_PERF_COUNT_SW_TASK_CLOCK

perf_type_id_PERF_TYPE_BREAKPOINT

perf_type_id_PERF_TYPE_HARDWARE

perf_type_id_PERF_TYPE_HW_CACHE

perf_type_id_PERF_TYPE_MAX

perf_type_id_PERF_TYPE_RAW

perf_type_id_PERF_TYPE_SOFTWARE

perf_type_id_PERF_TYPE_TRACEPOINT

Type Definitions

__be16

__be32

__be64

__kernel_caddr_t

__kernel_clock_t

__kernel_clockid_t

__kernel_daddr_t

__kernel_gid16_t

__kernel_gid32_t

__kernel_gid_t

__kernel_ino_t

__kernel_ipc_pid_t

__kernel_key_t

__kernel_loff_t

__kernel_long_t

__kernel_mode_t

__kernel_mqd_t

__kernel_off_t

__kernel_old_dev_t

__kernel_old_gid_t

__kernel_old_time_t

__kernel_old_uid_t

__kernel_pid_t

__kernel_ptrdiff_t

__kernel_sighandler_t

__kernel_size_t

__kernel_ssize_t

__kernel_suseconds_t

__kernel_time64_t

__kernel_time_t

__kernel_timer_t

__kernel_uid16_t

__kernel_uid32_t

__kernel_uid_t

__kernel_ulong_t

__le16

__le32

__le64

__poll_t

__s8

__s16

__s32

__s64

__sum16

__u8

__u16

__u32

__u64

__wsum

_bindgen_ty_1

_bindgen_ty_2

_bindgen_ty_3

_bindgen_ty_4

_bindgen_ty_5

bp_type_idx

perf_bpf_event_type

perf_branch_sample_type

perf_branch_sample_type_shift

perf_callchain_context

perf_event_ioc_flags

perf_event_ioctls

perf_event_read_format

perf_event_sample_format

perf_event_type

perf_hw_cache_id

perf_hw_cache_op_id

perf_hw_cache_op_result_id

perf_hw_id

perf_record_ksymbol_type

perf_sample_regs_abi

perf_sw_ids

perf_type_id

Unions

perf_event_attr__bindgen_ty_1

perf_event_attr__bindgen_ty_2

perf_event_attr__bindgen_ty_3

perf_event_attr__bindgen_ty_4

perf_event_mmap_page__bindgen_ty_1

perf_mem_data_src